voice-to-soap/requirements.txt
# AUDICIA VOICE-TO-SOAP PRODUCTION REQUIREMENTS
# HIPAA-Compliant Backend with Azure Integration

# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Azure Services
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-cognitiveservices-speech==1.32.1
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI Services
openai==1.3.5

# Security & Encryption
cryptography==41.0.8
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
structlog==23.2.0
httpx==0.25.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0
prometheus-client==0.19.0